Dirty hack to reset LLC state when the MS tries to reconnect with SGSN
A hack to get across the issue of mobile taking a lot of time to latch to the
network after some error forced it to reset its LLC send state variable.
This hack is a temporary change to let mobile latch to the network quickly.
The changes to configure.ac are also hackish in the interim and needs to be
cleaned-up later.